@@ -1,7 +1,7 @@
 {
   "name": "react-select",
   "version": "1.0.0-rc.3",
-  "description": "A Select control built with and for ReactJS",
+  "description": "A Select control built with and for ReactJS, forked for AdmiralCloud",
   "main": "lib/Select.js",
   "style": "dist/react-select.min.css",
   "author": "Jed Watson",
@@ -11,35 +11,35 @@
     "url": "https://github.com/JedWatson/react-select.git"
   },
   "dependencies": {
-    "classnames": "^2.2.4",
+    "classnames": "^2.2.5",
     "react-input-autosize": "^1.1.0"
   },
   "devDependencies": {
-    "babel": "^5.8.23",
-    "babel-eslint": "^4.1.3",
+    "babel": "^6.23.0",
+    "babel-eslint": "^7.2.1",
     "chai": "^3.5.0",
-    "coveralls": "^2.11.12",
-    "eslint": "^1.10.3",
-    "eslint-plugin-react": "^3.15.0",
+    "coveralls": "^2.12.0",
+    "eslint": "^3.18.0",
+    "eslint-plugin-react": "^6.10.3",
     "gulp": "^3.9.1",
     "isomorphic-fetch": "^2.2.1",
     "istanbul": "^0.4.5",
-    "jsdom": "^9.5.0",
-    "mocha": "^3.0.2",
-    "react": "^15.0",
-    "react-addons-shallow-compare": "^15.0",
-    "react-addons-test-utils": "^15.0",
+    "jsdom": "^9.12.0",
+    "mocha": "^3.2.0",
+    "react": "^15.4",
+    "react-addons-shallow-compare": "^15.4",
+    "react-addons-test-utils": "^15.4",
     "react-component-gulp-tasks": "^0.7.7",
-    "react-dom": "^15.0",
-    "react-gravatar": "^2.4.5",
-    "react-highlight-words": "^0.3.0",
-    "react-virtualized": "^7.23.0",
-    "react-virtualized-select": "^1.4.0",
-    "sinon": "^1.17.5",
-    "unexpected": "^10.17.0",
-    "unexpected-dom": "^3.1.0",
-    "unexpected-react": "^3.2.4",
-    "unexpected-sinon": "^10.4.0"
+    "react-dom": "^15.4",
+    "react-gravatar": "^2.6.1",
+    "react-highlight-words": "^0.6.0",
+    "react-virtualized": "^9.4.0",
+    "react-virtualized-select": "^2.4.0",
+    "sinon": "^2.1.0",
+    "unexpected": "^10.26.3",
+    "unexpected-dom": "^3.1.1",
+    "unexpected-react": "^3.5.2",
+    "unexpected-sinon": "^10.7.1"
   },
   "peerDependencies": {
     "react": "^0.14 || ^15.0.0-rc || ^15.0",
