@@ -8,22 +8,22 @@ const Value = React.createClass({
 	propTypes: {
 		children: React.PropTypes.node,
 		disabled: React.PropTypes.bool,               // disabled prop passed to ReactSelect
-		id: React.PropTypes.string,                   // Unique id for the value - used for aria
-		onClick: React.PropTypes.func,                // method to handle click on value label
+		id:       React.PropTypes.string,                   // Unique id for the value - used for aria
+		onClick:  React.PropTypes.func,                // method to handle click on value label
 		onRemove: React.PropTypes.func,               // method to handle removal of the value
-		value: React.PropTypes.object.isRequired,     // the option object for this value
+		value:    React.PropTypes.object.isRequired,     // the option object for this value
 	},
 
 	handleMouseDown (event) {
-		if (event.type === 'mousedown' && event.button !== 0) {
+		if(event.type === 'mousedown' && event.button !== 0) {
 			return;
 		}
-		if (this.props.onClick) {
+		if(this.props.onClick) {
 			event.stopPropagation();
 			this.props.onClick(this.props.value, event);
 			return;
 		}
-		if (this.props.value.href) {
+		if(this.props.value.href) {
 			event.stopPropagation();
 		}
 	},
@@ -54,14 +54,14 @@ const Value = React.createClass({
 	},
 
 	renderRemoveIcon () {
-		if (this.props.disabled || !this.props.onRemove) return;
+		if(this.props.disabled || !this.props.onRemove) return;
 		return (
-			<span className="Select-value-icon"
-				aria-hidden="true"
-				onMouseDown={this.onRemove}
-				onTouchEnd={this.handleTouchEndRemove}
-				onTouchStart={this.handleTouchStart}
-				onTouchMove={this.handleTouchMove}>
+				<span className="Select-value-icon"
+							aria-hidden="true"
+							onMouseDown={this.onRemove}
+							onTouchEnd={this.handleTouchEndRemove}
+							onTouchStart={this.handleTouchStart}
+							onTouchMove={this.handleTouchMove} >
 				&times;
 			</span>
 		);
@@ -70,11 +70,11 @@ const Value = React.createClass({
 	renderLabel () {
 		let className = 'Select-value-label';
 		return this.props.onClick || this.props.value.href ? (
-			<a className={className} href={this.props.value.href} target={this.props.value.target} onMouseDown={this.handleMouseDown} onTouchEnd={this.handleMouseDown}>
-				{this.props.children}
-			</a>
+				<a className={className} href={this.props.value.href} target={this.props.value.target} onMouseDown={this.handleMouseDown} onTouchEnd={this.handleMouseDown} >
+					{this.props.children}
+				</a>
 		) : (
-			<span className={className} role="option" aria-selected="true" id={this.props.id}>
+				<span className={className} role="option" aria-selected="true" id={this.props.id} >
 				{this.props.children}
 			</span>
 		);
@@ -82,13 +82,16 @@ const Value = React.createClass({
 
 	render () {
 		return (
-			<div className={classNames('Select-value', this.props.value.className)}
-				style={this.props.value.style}
-				title={this.props.value.title}
-				>
-				{this.renderRemoveIcon()}
-				{this.renderLabel()}
-			</div>
+				<div className={classNames('Select-value', this.props.value.className)}
+						 style={this.props.value.style}
+						 title={this.props.value.title} >
+					{
+						this.renderRemoveIcon()
+					}
+					{
+						this.renderLabel()
+					}
+				</div>
 		);
 	}
 
