@@ -1,8 +1,8 @@
 /*!
-  Copyright (c) 2016 Jed Watson.
-  Licensed under the MIT License (MIT), see
-  http://jedwatson.github.io/react-select
-*/
+ Copyright (c) 2016 Jed Watson.
+ Licensed under the MIT License (MIT), see
+ http://jedwatson.github.io/react-select
+ */
 
 'use strict';
 
@@ -69,6 +69,8 @@ var _Option2 = _interopRequireDefault(_Option);
 var _Value = require('./Value');
 
 var _Value2 = _interopRequireDefault(_Value);
+
+// AC change
 
 function stringifyValue(value) {
 	var valueType = typeof value;
@@ -160,9 +162,15 @@ var Select = _react2['default'].createClass({
 		valueComponent: _react2['default'].PropTypes.func, // value component to render
 		valueKey: _react2['default'].PropTypes.string, // path of the label value in option objects
 		valueRenderer: _react2['default'].PropTypes.func, // valueRenderer: function (option) {}
-		wrapperStyle: _react2['default'].PropTypes.object },
+		wrapperStyle: _react2['default'].PropTypes.object, // optional style to apply to the component wrapper
+		//
+		// AC change
+		allowCreate: _react2['default'].PropTypes.bool, // whether to allow creation of new entries
+		multiSelectListBelow: _react2['default'].PropTypes.bool, // render list of selected items below the selectthis field name for html forms
+		newOptionCreator: _react2['default'].PropTypes.func, // factory to create new options when allowCreate set
+		showSelectedCount: _react2['default'].PropTypes.bool },
 
-	// optional style to apply to the component wrapper
+	// show selected item count instead of default
 	statics: { Async: _Async2['default'], AsyncCreatable: _AsyncCreatable2['default'], Creatable: _Creatable2['default'] },
 
 	getDefaultProps: function getDefaultProps() {
@@ -205,7 +213,13 @@ var Select = _react2['default'].createClass({
 			simpleValue: false,
 			tabSelectsValue: true,
 			valueComponent: _Value2['default'],
-			valueKey: 'value'
+			valueKey: 'value',
+			//
+			// AC change
+			allowCreate: false,
+			multiSelectListBelow: false,
+			resetValue: null,
+			showSelectedCount: false
 		};
 	},
 
@@ -215,7 +229,10 @@ var Select = _react2['default'].createClass({
 			isFocused: false,
 			isOpen: false,
 			isPseudoFocused: false,
-			required: false
+			required: false,
+			//
+			// AC change
+			isLoading: false
 		};
 	},
 
@@ -496,6 +513,12 @@ var Select = _react2['default'].createClass({
 	handleInputChange: function handleInputChange(event) {
 		var newInputValue = event.target.value;
 
+		console.log('xxx handleInputChange', newInputValue, 'this.state.inputValue', this.state.inputValue);
+		// AC specific Fix IE11 Bug
+		if (this.state.inputValue === '' && newInputValue === '') {
+			return;
+		}
+
 		if (this.state.inputValue !== event.target.value && this.props.onInputChange) {
 			var nextState = this.props.onInputChange(newInputValue);
 			// Note: != used deliberately here to catch undefined and null
@@ -589,6 +612,15 @@ var Select = _react2['default'].createClass({
 					this.popValue();
 				}
 				return;
+			// case 188: // AC change
+			// 	if (this.props.allowCreate && this.props.multi) {
+			// 		event.preventDefault();
+			// 		event.stopPropagation();
+			// 		this.selectFocusedOption();
+			// 	} else {
+			// 		return;
+			// 	}
+			// break;
 			default:
 				return;
 		}
@@ -620,9 +652,9 @@ var Select = _react2['default'].createClass({
 
 	/**
   * Turns a value into an array from the given options
-  * @param	{String|Number|Array}	value		- the value of the select input
-  * @param	{Object}		nextProps	- optionally specify the nextProps so the returned array uses the latest configuration
-  * @returns	{Array}	the value of the select represented in an array
+  * @param  {String|Number|Array}  value    - the value of the select input
+  * @param  {Object}    nextProps  - optionally specify the nextProps so the returned array uses the latest configuration
+  * @returns  {Array}  the value of the select represented in an array
   */
 	getValueArray: function getValueArray(value, nextProps) {
 		var _this = this;
@@ -647,8 +679,8 @@ var Select = _react2['default'].createClass({
 
 	/**
   * Retrieve a value from the given options and valueKey
-  * @param	{String|Number|Array}	value	- the selected value(s)
-  * @param	{Object}		props	- the Select component's props (or nextProps)
+  * @param  {String|Number|Array}  value  - the selected value(s)
+  * @param  {Object}    props  - the Select component's props (or nextProps)
   */
 	expandValue: function expandValue(value, props) {
 		var valueType = typeof value;
@@ -879,13 +911,11 @@ var Select = _react2['default'].createClass({
 
 		var renderLabel = this.props.valueRenderer || this.getOptionLabel;
 		var ValueComponent = this.props.valueComponent;
-		if (!valueArray.length) {
-			return !this.state.inputValue ? _react2['default'].createElement(
-				'div',
-				{ className: 'Select-placeholder' },
-				this.props.placeholder
-			) : null;
-		}
+		// AC change
+		console.log('xxx valueArray.length -> show placeholder', !valueArray.length);
+		// if(!valueArray.length) {
+		// 	return !this.state.inputValue ? <div className="Select-placeholder" >{this.props.placeholder}</div> : null;
+		// }
 		var onClick = this.props.onValueClick ? this.handleValueClick : null;
 		if (this.props.multi) {
 			return valueArray.map(function (value, i) {
@@ -898,8 +928,7 @@ var Select = _react2['default'].createClass({
 						key: 'value-' + i + '-' + value[_this4.props.valueKey],
 						onClick: onClick,
 						onRemove: _this4.removeValue,
-						value: value
-					},
+						value: value },
 					renderLabel(value, i),
 					_react2['default'].createElement(
 						'span',
@@ -917,16 +946,39 @@ var Select = _react2['default'].createClass({
 					disabled: this.props.disabled,
 					instancePrefix: this._instancePrefix,
 					onClick: onClick,
-					value: valueArray[0]
-				},
+					value: valueArray[0] },
 				renderLabel(valueArray[0])
 			);
 		}
 	},
 
+	// AC change
+	renderMultiSelectedList: function renderMultiSelectedList(valueArray) {
+		var _this5 = this;
+
+		var renderLabel = this.props.valueRenderer || this.getOptionLabel;
+		var onClick = this.props.onValueClick ? this.handleValueClick : null;
+
+		return valueArray.map(function (value, i) {
+			return _react2['default'].createElement(
+				MultiSelectValueList,
+				{
+					id: _this5._instancePrefix + '-value-' + i,
+					instancePrefix: _this5._instancePrefix,
+					disabled: _this5.props.disabled || value.clearableValue === false,
+					key: 'value-' + i + '-' + value[_this5.props.valueKey],
+					onClick: onClick,
+					onRemove: _this5.removeValue,
+					value: value,
+					className: valueArray.length === 1 ? 'single' : '' },
+				renderLabel(value)
+			);
+		});
+	},
+
 	renderInput: function renderInput(valueArray, focusedOptionIndex) {
 		var _classNames,
-		    _this5 = this;
+		    _this6 = this;
 
 		var className = (0, _classnames2['default'])('Select-input', this.props.inputProps.className);
 		var isOpen = !!this.state.isOpen;
@@ -948,12 +1000,13 @@ var Select = _react2['default'].createClass({
 			onChange: this.handleInputChange,
 			onFocus: this.handleInputFocus,
 			ref: function ref(_ref) {
-				return _this5.input = _ref;
+				return _this6.input = _ref;
 			},
 			required: this.state.required,
-			value: this.state.inputValue
-		});
+			value: this.state.inputValue, // AC change
+			placeholder: this.props.placeholder });
 
+		// AC change
 		if (this.props.inputRenderer) {
 			return this.props.inputRenderer(inputProps);
 		}
@@ -974,7 +1027,7 @@ var Select = _react2['default'].createClass({
 				onBlur: this.handleInputBlur,
 				onFocus: this.handleInputFocus,
 				ref: function (ref) {
-					return _this5.input = ref;
+					return _this6.input = ref;
 				},
 				'aria-readonly': '' + !!this.props.disabled,
 				style: { border: 0, width: 1, display: 'inline-block' } }));
@@ -1025,7 +1078,11 @@ var Select = _react2['default'].createClass({
 	filterOptions: function filterOptions(excludeOptions) {
 		var filterValue = this.state.inputValue;
 		var options = this.props.options || [];
-		if (this.props.filterOptions) {
+
+		// AC Adjustment special setting for multi select lists
+		if (this.props.filterOptions && _.isEmpty(filterValue) && !this.props.multiSelectListBelow) {
+			return options;
+		} else if (this.props.filterOptions) {
 			// Maintain backwards compatibility with boolean attribute
 			var filterOptions = typeof this.props.filterOptions === 'function' ? this.props.filterOptions : _utilsDefaultFilterOptions2['default'];
 
@@ -1079,17 +1136,17 @@ var Select = _react2['default'].createClass({
 	},
 
 	renderHiddenField: function renderHiddenField(valueArray) {
-		var _this6 = this;
+		var _this7 = this;
 
 		if (!this.props.name) return;
 		if (this.props.joinValues) {
 			var value = valueArray.map(function (i) {
-				return stringifyValue(i[_this6.props.valueKey]);
+				return stringifyValue(i[_this7.props.valueKey]);
 			}).join(this.props.delimiter);
 			return _react2['default'].createElement('input', {
 				type: 'hidden',
 				ref: function (ref) {
-					return _this6.value = ref;
+					return _this7.value = ref;
 				},
 				name: this.props.name,
 				value: value,
@@ -1099,9 +1156,9 @@ var Select = _react2['default'].createClass({
 			return _react2['default'].createElement('input', { key: 'hidden.' + index,
 				type: 'hidden',
 				ref: 'value' + index,
-				name: _this6.props.name,
-				value: stringifyValue(item[_this6.props.valueKey]),
-				disabled: _this6.props.disabled });
+				name: _this7.props.name,
+				value: stringifyValue(item[_this7.props.valueKey]),
+				disabled: _this7.props.disabled });
 		});
 	},
 
@@ -1111,7 +1168,14 @@ var Select = _react2['default'].createClass({
 
 		var focusedOption = this.state.focusedOption || selectedOption;
 		if (focusedOption && !focusedOption.disabled) {
-			var focusedOptionIndex = options.indexOf(focusedOption);
+			var focusedOptionIndex = -1;
+			options.some(function (option, index) {
+				var isOptionEqual = option.value === focusedOption.value;
+				if (isOptionEqual) {
+					focusedOptionIndex = index;
+				}
+				return isOptionEqual;
+			});
 			if (focusedOptionIndex !== -1) {
 				return focusedOptionIndex;
 			}
@@ -1124,7 +1188,7 @@ var Select = _react2['default'].createClass({
 	},
 
 	renderOuter: function renderOuter(options, valueArray, focusedOption) {
-		var _this7 = this;
+		var _this8 = this;
 
 		var menu = this.renderMenu(options, valueArray, focusedOption);
 		if (!menu) {
@@ -1134,12 +1198,12 @@ var Select = _react2['default'].createClass({
 		return _react2['default'].createElement(
 			'div',
 			{ ref: function (ref) {
-					return _this7.menuContainer = ref;
+					return _this8.menuContainer = ref;
 				}, className: 'Select-menu-outer', style: this.props.menuContainerStyle },
 			_react2['default'].createElement(
 				'div',
 				{ ref: function (ref) {
-						return _this7.menu = ref;
+						return _this8.menu = ref;
 					}, role: 'listbox', className: 'Select-menu', id: this._instancePrefix + '-list',
 					style: this.props.menuStyle,
 					onScroll: this.handleMenuScroll,
@@ -1150,7 +1214,7 @@ var Select = _react2['default'].createClass({
 	},
 
 	render: function render() {
-		var _this8 = this;
+		var _this9 = this;
 
 		var valueArray = this.getValueArray(this.props.value);
 		var options = this._visibleOptions = this.filterOptions(this.props.multi ? this.getValueArray(this.props.value) : null);
@@ -1185,10 +1249,15 @@ var Select = _react2['default'].createClass({
 			);
 		}
 
+		// AC change
+		var multiSelectListStyle = (0, _classnames2['default'])({
+			'Select--ItemsWrap': this.props.multiSelectListBelow && valueArray.length > 0
+		});
+
 		return _react2['default'].createElement(
 			'div',
 			{ ref: function (ref) {
-					return _this8.wrapper = ref;
+					return _this9.wrapper = ref;
 				},
 				className: className,
 				style: this.props.wrapperStyle },
@@ -1196,7 +1265,7 @@ var Select = _react2['default'].createClass({
 			_react2['default'].createElement(
 				'div',
 				{ ref: function (ref) {
-						return _this8.control = ref;
+						return _this9.control = ref;
 					},
 					className: 'Select-control',
 					style: this.props.style,
@@ -1204,12 +1273,12 @@ var Select = _react2['default'].createClass({
 					onMouseDown: this.handleMouseDown,
 					onTouchEnd: this.handleTouchEnd,
 					onTouchStart: this.handleTouchStart,
-					onTouchMove: this.handleTouchMove
-				},
+					onTouchMove: this.handleTouchMove },
 				_react2['default'].createElement(
 					'span',
 					{ className: 'Select-multi-value-wrapper', id: this._instancePrefix + '-value' },
-					this.renderValue(valueArray, isOpen),
+					 // AC change
+					!this.props.showSelectedCount && !this.props.multiSelectListBelow && !!this.props.filterOptions && valueArray.length > 0 ? this.renderValue(valueArray, isOpen) : null,
 					this.renderInput(valueArray, focusedOptionIndex)
 				),
 				removeMessage,
@@ -1217,7 +1286,13 @@ var Select = _react2['default'].createClass({
 				this.renderClear(),
 				this.renderArrow()
 			),
-			isOpen ? this.renderOuter(options, !this.props.multi ? valueArray : null, focusedOption) : null
+			isOpen ? this.renderOuter(options, !this.props.multi ? valueArray : null, focusedOption) : null,
+			 // AC change
+			this.props.multiSelectListBelow && valueArray.length > 0 ? _react2['default'].createElement(
+				'div',
+				{ className: multiSelectListStyle },
+				this.renderMultiSelectedList(valueArray)
+			) : null
 		);
 	}
 
