@@ -94,8 +94,7 @@ var Value = _react2['default'].createClass({
 			'div',
 			{ className: (0, _classnames2['default'])('Select-value', this.props.value.className),
 				style: this.props.value.style,
-				title: this.props.value.title
-			},
+				title: this.props.value.title },
 			this.renderRemoveIcon(),
 			this.renderLabel()
 		);
