@@ -14,8 +14,7 @@ var _react2 = _interopRequireDefault(_react);
 function clearRenderer() {
 	return _react2['default'].createElement('span', {
 		className: 'Select-clear',
-		dangerouslySetInnerHTML: { __html: '&times;' }
-	});
+		dangerouslySetInnerHTML: { __html: '&times;' } });
 }
 
 ;
