@@ -51,8 +51,7 @@ function menuRenderer(_ref) {
 				optionIndex: i,
 				ref: function (ref) {
 					onOptionRef(ref, isFocused);
-				}
-			},
+				} },
 			optionRenderer(option, i)
 		);
 	});
