@@ -1,13 +1,13 @@
 import React from 'react';
 import Select from './Select';
 
-function reduce(obj, props = {}){
-  return Object.keys(obj)
-  .reduce((props, key) => {
-    const value = obj[key];
-    if (value !== undefined) props[key] = value;
-    return props;
-  }, props);
+function reduce(obj, props = {}) {
+	return Object.keys(obj)
+			.reduce((props, key) => {
+				const value = obj[key];
+				if(value !== undefined) props[key] = value;
+				return props;
+			}, props);
 }
 
 const AsyncCreatable = React.createClass({
@@ -15,25 +15,25 @@ const AsyncCreatable = React.createClass({
 
 	render () {
 		return (
-			<Select.Async {...this.props}>
-				{(asyncProps) => (
-					<Select.Creatable {...this.props}>
-						{(creatableProps) => (
-							<Select
-								{...reduce(asyncProps, reduce(creatableProps, {}))}
-								onInputChange={(input) => {
-									creatableProps.onInputChange(input);
-									return asyncProps.onInputChange(input);
-								}}
-								ref={(ref) => {
-									creatableProps.ref(ref);
-									asyncProps.ref(ref);
-								}}
-							/>
-						)}
-					</Select.Creatable>
-				)}
-			</Select.Async>
+				<Select.Async {...this.props}>
+					{(asyncProps) => (
+							<Select.Creatable {...this.props}>
+								{(creatableProps) => (
+										<Select
+												{...reduce(asyncProps, reduce(creatableProps, {}))}
+												onInputChange={(input) => {
+													creatableProps.onInputChange(input);
+													return asyncProps.onInputChange(input);
+												}}
+												ref={(ref) => {
+													creatableProps.ref(ref);
+													asyncProps.ref(ref);
+												}}
+										/>
+								)}
+							</Select.Creatable>
+					)}
+				</Select.Async>
 		);
 	}
 });
