@@ -4,7 +4,6 @@ export default function clearRenderer () {
 	return (
 		<span
 			className="Select-clear"
-			dangerouslySetInnerHTML={{ __html: '&times;' }}
-		/>
+			dangerouslySetInnerHTML={{ __html: '&times;' }} />
 	);
 };
